:recycle: crawler 모듈을 js -> ts 로 클래스화
- earthstar
- goldenkey
- keyescape
- masterkey
- nextedition
- secretgardenescape
- seoulescape
- xdungeon
